Refactoring to reduce Sketchup dependency
