+++ .gitignore
@@ -5,3 +5,4 @@ launch.json
 **/*~
 .yardoc
 doc/
+coverage/
