+++ Makefile
@@ -1,13 +1,18 @@
 VERSION=2.3.0
 plugin: check
-	rm -f facesvg*.rbz; cd lib; zip -r ../facesvg-$(VERSION).rbz facesvg.rb facesvg/*rb
+	rm -f facesvg*.rbz;
+	cd lib; zip -r ../facesvg-$(VERSION).rbz $(find . -name '*rb' -o -name '*png')
 	# CONVENIENCE ONLY
-	-mkdir \
-	   ~/'Library/Application Support/SketchUp 2022/SketchUp/Plugins/facesvg/'
+	rm -rf ~/'Library/Application Support/SketchUp 2022/SketchUp/Plugins/facesvg/'
 	cp lib/facesvg.rb \
 	   ~/'Library/Application Support/SketchUp 2022/SketchUp/Plugins'
-	cp -f lib/facesvg/*rb \
-	   ~/'Library/Application Support/SketchUp 2022/SketchUp/Plugins/facesvg/'
+	cp -pr lib/facesvg \
+	   ~/'Library/Application Support/SketchUp 2022/SketchUp/Plugins/facesvg'
 
 check:
-	rubocop lib
+	rubocop lib test
+
+test:
+	ruby run_test.rb
+
+.PHONY: test check
